refactor: 유저 롤, 도메인명 변경 Owner -> Seller
