--- src/main/java/com/example/jjangushrine/domain/owner/controller/OwnerController.java
@@ -1,7 +1,0 @@
-package com.example.jjangushrine.domain.owner.controller;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class OwnerController {
-}
