--- src/main/java/com/example/jjangushrine/domain/owner/dto/OwnerValidation.java
@@ -1,4 +1,0 @@
-package com.example.jjangushrine.domain.owner.dto;
-
-public class OwnerValidation {
-}
