--- src/main/java/com/example/jjangushrine/domain/owner/dto/request/ownerReq.java
@@ -1,4 +1,0 @@
-package com.example.jjangushrine.domain.owner.dto.request;
-
-public class ownerReq {
-}
