--- src/main/java/com/example/jjangushrine/domain/owner/dto/response/OwnerRes.java
@@ -1,4 +1,0 @@
-package com.example.jjangushrine.domain.owner.dto.response;
-
-public class OwnerRes {
-}
