--- src/main/java/com/example/jjangushrine/domain/owner/entity/Owner.java
@@ -1,73 +1,0 @@
-package com.example.jjangushrine.domain.owner.entity;
-
-import java.time.LocalDateTime;
-
-import org.hibernate.annotations.ColumnDefault;
-
-import com.example.jjangushrine.domain.user.enums.UserRole;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "owners")
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class Owner {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "owner_id")
-	private Long id;
-
-	@Column(nullable = false)
-	private String email;
-
-	@Column(nullable = false)
-	private String password;
-
-	@Column(nullable = false)
-	private String representativeName;
-
-	@Column(nullable = false)
-	private String phoneNumber;
-
-	@Column(nullable = false)
-	private String address;
-
-	@Column(nullable = false)
-	private String addressDetail;
-
-	@Column(nullable = false)
-	private String zipCode;
-
-	@Builder.Default
-	@Column(nullable = false)
-	@Enumerated(EnumType.STRING)
-	private UserRole userRole = UserRole.OWNER;
-
-	@Column
-	@ColumnDefault("false")
-	private Boolean isDeleted;
-
-	@Column
-	private LocalDateTime deletedAt;
-
-	public void softDelete() {
-		this.isDeleted = true;
-		this.deletedAt = LocalDateTime.now();
-	}
-}
