--- src/main/java/com/example/jjangushrine/domain/owner/repository/OwnerRepository.java
@@ -1,4 +1,0 @@
-package com.example.jjangushrine.domain.owner.repository;
-
-public class OwnerRepository {
-}
