--- src/main/java/com/example/jjangushrine/domain/owner/service/OwnerService.java
@@ -1,4 +1,0 @@
-package com.example.jjangushrine.domain.owner.service;
-
-public class OwnerService {
-}
