+++ src/main/java/com/example/jjangushrine/domain/seller/controller/SellerController.java
@@ -1,0 +1,7 @@
+package com.example.jjangushrine.domain.seller.controller;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class SellerController {
+}
