+++ src/main/java/com/example/jjangushrine/domain/seller/dto/SellerValidation.java
@@ -1,0 +1,4 @@
+package com.example.jjangushrine.domain.seller.dto;
+
+public class SellerValidation {
+}
