+++ src/main/java/com/example/jjangushrine/domain/seller/dto/request/SellerReq.java
@@ -1,0 +1,4 @@
+package com.example.jjangushrine.domain.seller.dto.request;
+
+public class SellerReq {
+}
