+++ src/main/java/com/example/jjangushrine/domain/seller/dto/response/SellerRes.java
@@ -1,0 +1,4 @@
+package com.example.jjangushrine.domain.seller.dto.response;
+
+public class SellerRes {
+}
