+++ src/main/java/com/example/jjangushrine/domain/seller/repository/SellerRepository.java
@@ -1,0 +1,4 @@
+package com.example.jjangushrine.domain.seller.repository;
+
+public class SellerRepository {
+}
