+++ src/main/java/com/example/jjangushrine/domain/seller/service/SellerService.java
@@ -1,0 +1,4 @@
+package com.example.jjangushrine.domain.seller.service;
+
+public class SellerService {
+}
