+++ src/main/java/com/example/jjangushrine/domain/user/enums/UserRole.java
@@ -5,12 +5,12 @@ import java.util.Arrays;
 import com.example.jjangushrine.exception.common.InvalidException;
 
 public enum UserRole {
-	OWNER, USER;
+	SELLER, USER;
 
 	public static UserRole of(String role) {
 		return Arrays.stream(UserRole.values())
 			.filter(r -> r.name().equalsIgnoreCase(role))
 			.findFirst()
-			.orElseThrow(()-> new InvalidException());
+			.orElseThrow(InvalidException::new);
 	}
 }
